Moved and updated Task
